Refs #192 - Get filter callback the new way.
Abstract out shared test methods.
