+++ src/Synapse/Validator/Constraints/RowExistsValidator.php
@@ -18,8 +18,8 @@ class RowExistsValidator extends ConstraintValidator
      */
     public function validate($value, Constraint $constraint)
     {
-        $callback = $constraint->filterCallback;
-        $entity = $constraint->getMapper()->findBy($callback($value));
+        $callback = $constraint->getFilterCallback();
+        $entity   = $constraint->getMapper()->findBy($callback($value));
 
         if ($entity instanceof AbstractEntity) {
             return;
